Inventário e Resolução
-Sistema de inventário melhorado.
-Resolução ajustada
