+++ OneFish/Assets/Resources/Items/Fish1.asset
@@ -13,5 +13,4 @@ MonoBehaviour:
   m_Name: Fish1
   m_EditorClassIdentifier: 
   itemName: Salmon
-  itemSprite: {fileID: 21300000, guid: a2942f74ee121044086f8c6c07451c76, type: 3}
-  itemTag: 0
+  icon: {fileID: 21300000, guid: a2942f74ee121044086f8c6c07451c76, type: 3}
