+++ OneFish/Assets/Resources/Items/Fish2.asset
@@ -13,5 +13,4 @@ MonoBehaviour:
   m_Name: Fish2
   m_EditorClassIdentifier: 
   itemName: tuna
-  itemSprite: {fileID: 21300000, guid: 99a8704b9f57f5b4fb8b9d9fdafe3fa5, type: 3}
-  itemTag: 0
+  icon: {fileID: 21300000, guid: 99a8704b9f57f5b4fb8b9d9fdafe3fa5, type: 3}
