+++ OneFish/Assets/Scripts/Inventory/Bar/BagScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d71a491b93806ac4299a5099a2c7c970
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
