--- OneFish/Assets/Scripts/Inventory/Inventory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2f56f520af38d2845aec8c26a04c851c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
