+++ OneFish/Assets/Scripts/Inventory/Slot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b1781d6e8d12a343b3da83061259a16
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
