--- OneFish/Assets/Scripts/MenuMeneger.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 02371c7920a256747bd04d4ba905be3e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
