+++ OneFish/Assets/Scripts/Models/Item.cs
@@ -1,11 +1,10 @@
 using UnityEngine;
 
 public enum SlotTag { Peixe, None }
-[CreateAssetMenu(menuName = "OneFish/Item")]
+[CreateAssetMenu(fileName = "NewItem", menuName = "OneFish/Item")]
 public class Item : ScriptableObject
 {
-    public string itemName;
-    public Sprite itemSprite;
-    public SlotTag itemTag;
+    public string itemName; 
+    public Sprite icon; 
 
 }
