+++ OneFish/Assets/Scripts/Player/PlayerControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d7df6cfbd826f5a4c8af6349537e9ec1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
