--- OneFish/Assets/Scripts/Teste.cs
@@ -1,12 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Teste : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-         Screen.SetResolution(1920, 1080, FullScreenMode.Windowed);
-    }
-}
