--- OneFish/Assets/Scripts/Teste.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d03cc134227242f43a475f4cb0fd69c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
