+++ OneFish/ProjectSettings/TagManager.asset
@@ -14,7 +14,7 @@ TagManager:
   - FishArea
   - Player
   - Barrier
-  - 
+  - Inv
   - 
   - 
   - 
@@ -43,4 +43,7 @@ TagManager:
     locked: 0
   - name: FishingBar
     uniqueID: 3569408921
+    locked: 0
+  - name: Inventory
+    uniqueID: 1506638181
     locked: 0
