+++ OneFish/UserSettings/EditorUserSettings.asset
@@ -6,10 +6,10 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 0202525355540a0e0f5a5c7646200844144e4d732a2a22322e2f4463b3e1603a
+      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
       flags: 0
     RecentlyUsedSceneGuid-1:
-      value: 515250075c0c595e5f5a5e71122159444e4e4a2f7a7d7f602f284d66b4b76661
+      value: 0202525355540a0e0f5a5c7646200844144e4d732a2a22322e2f4463b3e1603a
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
